# swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Lokal.ind API",
    "description": "Documentation of Lokal.ind API",
    "contact": {
      "name": "C241-PS127 Cloud Computing team",
      "url": "https://github.com/C241-PS127/Cloud-Computing-API"
    },
    "version": "1.0.0"
  },
  "host": "lokal-ind.et.r.appspot.com/",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/products/": {
      "get": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productName": {
                  "example": "any"
                },
                "productDescription": {
                  "example": "any"
                },
                "brandId": {
                  "example": "any"
                },
                "categoryId": {
                  "example": "any"
                },
                "unitPrice": {
                  "example": "any"
                },
                "unitSize": {
                  "example": "any"
                },
                "unitInStock": {
                  "example": "any"
                },
                "isAvailable": {
                  "example": "any"
                },
                "rating": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{ProductId}": {
      "get": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "ProductId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "ProductId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productName": {
                  "example": "any"
                },
                "productDescription": {
                  "example": "any"
                },
                "categoryId": {
                  "example": "any"
                },
                "unitPrice": {
                  "example": "any"
                },
                "unitSize": {
                  "example": "any"
                },
                "unitInStock": {
                  "example": "any"
                },
                "isAvailable": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "ProductId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/brand/{BrandName}": {
      "get": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "BrandName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/category/{CategoryName}": {
      "get": {
        "tags": ["Products"],
        "description": "",
        "parameters": [
          {
            "name": "CategoryName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/wishlist/": {
      "post": {
        "tags": ["Wishlist"],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Wishlist"],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/wishlist/myWishlist": {
      "get": {
        "tags": ["Wishlist"],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/wishlist/{wishListId}": {
      "delete": {
        "tags": ["Wishlist"],
        "description": "",
        "parameters": [
          {
            "name": "wishListId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}